requirements.txt
torch
numpy
scipy
matplotlib
seaborn
pyyaml








